# docs/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "MAIN Smart Parking",
    "version": "1.0.0",
    "description": "SRS 2.3. ADMIN includes USER privileges; OWNER means resource ownership or ADMIN. Business ALLOW/DENY both return 200. Timestamps UTC; history ranges [from,to)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/cards/{id}/renewal-orders": {
      "post": {
        "operationId": "post__cards_id_renewal-orders",
        "tags": [
          "OWNER"
        ],
        "summary": "POST /cards/{id}/renewal-orders",
        "security": [
          {
            "Session": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain fresh context after login/logout via /auth/csrf."
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        }
      }
    },
...
    "/renewal-orders/{id}": {
      "get": {
        "operationId": "get__renewal-orders_id",
        "tags": [
          "OWNER"
        ],
        "summary": "GET /renewal-orders/{id}",
        "security": [
          {
            "Session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/renewal-orders/{id}/reconcile": {
      "post": {
        "operationId": "post__renewal-orders_id_reconcile",
        "tags": [
          "OWNER"
        ],
        "summary": "POST /renewal-orders/{id}/reconcile",
        "security": [
          {
            "Session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain fresh context after login/logout via /auth/csrf."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileRequest"
              }
            }
          }
        }
      }
    },
...
    "/payments/webhooks/payos": {
      "post": {
        "operationId": "post__payments_webhooks_payos",
        "tags": [
          "WEBHOOK"
        ],
        "summary": "POST /payments/webhooks/payos",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayosWebhook"
              }
            }
          }
        },
        "description": "payOS signed data object required. Session/CSRF cannot authorize payment. Route rejects in mock mode."
      }
    },
...
  },
  "components": {
    "schemas": {
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "fieldErrors",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "ALREADY_AUTHENTICATED",
              "ALREADY_INSIDE",
              "AUTH_REQUIRED",
              "BODY_TOO_LARGE",
              "CARD_BLOCKED",
              "CSRF_INVALID",
              "CURRENT_PASSWORD_INVALID",
              "DATABASE_UNAVAILABLE",
              "DEVICE_BOOT_NOT_REGISTERED",
              "DEVICE_DISABLED",
              "DEVICE_RATE_LIMITED",
              "DEVICE_UNAUTHORIZED",
              "ENABLED_CARD_EXISTS",
              "FORBIDDEN",
              "IDEMPOTENCY_CONFLICT",
              "INVALID_CREDENTIALS",
              "INVALID_SESSION_STATE",
              "INVALID_SIGNATURE",
              "LOGIN_RATE_LIMITED",
              "NOT_FOUND",
              "ORDER_IN_PROGRESS",
              "PACKAGE_DISABLED",
              "PAYMENT_CONSISTENCY_ERROR",
              "PAYMENT_MODE_MISMATCH",
              "PAYMENT_PROVIDER_UNAVAILABLE",
              "PROVIDER_MISMATCH",
              "RECONCILE_RATE_LIMITED",
              "REGISTRATION_RATE_LIMITED",
              "SESSION_CAPACITY_CONFLICT",
              "SESSION_INVALID",
              "UID_TAKEN",
              "USERNAME_TAKEN",
              "USER_DISABLED",
              "VALIDATION_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "traceId": {
            "type": "string"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "description": "Existing order for ORDER_IN_PROGRESS; only returned within authorized ownership."
          }
        }
      },
...
      "OrderView": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "cardId",
          "packageCode",
          "durationDays",
          "amountVnd",
          "currency",
          "paymentMode",
          "status",
          "createdAt",
          "expiresAt",
          "serverTime",
          "qrPayload",
          "checkoutUrl",
          "paidAt",
          "renewalAppliedAt",
          "oldExpiresAt",
          "newExpiresAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "packageCode": {
            "type": "string"
          },
          "durationDays": {
            "type": "integer"
          },
          "amountVnd": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "paymentMode": {
            "type": "string",
            "enum": [
              "mock",
              "payos"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATING",
              "PENDING",
              "PAID",
              "EXPIRED",
              "FAILED",
              "REVIEW"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          },
          "qrPayload": {
            "type": "string",
            "nullable": true
          },
          "checkoutUrl": {
            "type": "string",
            "nullable": true
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "renewalAppliedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "oldExpiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "newExpiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "packageId"
        ],
        "properties": {
          "packageId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ReconcileRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "acceptLatePayment"
        ],
        "properties": {
          "acceptLatePayment": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "WebhookData": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "orderCode",
          "amount",
          "paymentLinkId",
          "reference",
          "currency",
          "transactionDateTime",
          "code",
          "desc",
          "description",
          "accountNumber"
        ],
        "properties": {
          "orderCode": {
            "type": "integer"
          },
          "amount": {
            "type": "number"
          },
          "paymentLinkId": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "transactionDateTime": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "counterAccountBankId": {
            "type": "string"
          },
          "counterAccountBankName": {
            "type": "string"
          },
          "counterAccountName": {
            "type": "string"
          },
          "counterAccountNumber": {
            "type": "string"
          },
          "virtualAccountName": {
            "type": "string"
          },
          "virtualAccountNumber": {
            "type": "string"
          }
        }
      },
      "PayosWebhook": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "desc",
          "success",
          "data",
          "signature"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/WebhookData"
          },
          "signature": {
            "type": "string"
          }
        }
      },
      "Ack": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accepted"
        ],
        "properties": {
          "accepted": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      },
      "DeviceId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Id"
      },
      "DeviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Key"
      }
    }
  }
}